+++ src/ledger/model/Currency.cpp
@@ -14,6 +14,9 @@
  * limitations under the License.
  */
 
+// Standard include(s)
+#include <cstdint>
+
 // Qt include(s)
 #include <QtCore>
 
@@ -226,7 +229,7 @@ const QMap<QString, QString> & symbols() {
 Currency::Currency() : id_(1), iso4217_(QLocale().currencySymbol(QLocale::CurrencyIsoCode)) {}
 
 //--------------------------------------------------------------------------------------------------
-Currency::Currency(int id, const QString & code) : id_(id), iso4217_(code) {}
+Currency::Currency(int64_t id, const QString & code) : id_(id), iso4217_(code) {}
 
 //--------------------------------------------------------------------------------------------------
 bool Currency::operator==(const Currency & that) const {
@@ -249,7 +252,7 @@ QString Currency::format(double value) const {
 }
 
 //--------------------------------------------------------------------------------------------------
-int Currency::id() const {
+int64_t Currency::id() const {
     return id_;
 }
